window_sdl2: fix input_type for Android Keyboards
Since the SDL2 adoption, it is not possible to use the input_type
feature in TextInput to change the displayed keyboard in Android. The
reason is that it has not been implemented yet.

In order to make this change possible, the "android.show_keyboard"
function is used, which through the JNI calls "showKeyboard", located in
"org/renpy/Hardware.java".

As "android.show_keyboard" requires an input type value, it is provided
to the "_WindowSDL2Storage.show_keyboard" method under the "input_type"
parameter.